Fix TIM compatibility and improve TIM validation.
PSn00bSDK is a good SDK for testing decompiled game code and working with extracted TIM resources.
However, there is an issue where some textures that can be loaded with the PsyQ SDK
are not compatible with PSn00bSDK, causing a black screen when attempting to load them.
This commit fixes this issue.

Additionally, it adds a new check in the `GsGetTimInfo` and `GetTimInfo` functions.
The check verifies if pixel data exists before setting the corresponding fields in the structure,
preventing the loading of garbage data.